+++ hicn-light/src/hicn/core/nameBitvector.c
@@ -240,7 +240,7 @@ uint32_t nameBitvector_lpm(const NameBitvector *a,
 
   uint64_t diff = a->bits[0] ^ b->bits[0];
   if(diff){
-    prefix_len = BV_SIZE - (_diff_bit_log2(diff) + 1);
+    prefix_len = (uint32_t)(BV_SIZE - (_diff_bit_log2(diff) + 1));
     //printf("if 1 diff = %lu plen = %d\n", diff, prefix_len);
   }else{
     prefix_len = BV_SIZE;
